change server

+++ server-side.c
@@ -21,8 +21,8 @@ float delay = 0;
 // WARNING: the packet size is given in bytes so a conversion to bits is needed
 float link_delay_tx(int packet_size) {
 	printf("Received packet size: %i\n", packet_size);
-	float delay = (packet_size * 8)/LINK_SPEED;
-	printf("Generated delay: %f for packet size %f\n", delay, packet_size);
+	double delay = (packet_size * 8)/LINK_SPEED;
+	printf("Generated delay: %f for packet size %lf\n", delay, packet_size);
 	return delay;
 }
